RecordBrowser: get_records_limit() renamed to get_records_count() integer and float are now two different types style "number" applied to numeric fields

+++ modules/Data/TaxRates/TaxRatesInstall.php
@@ -17,7 +17,7 @@ class Data_TaxRatesInstall extends ModuleInstall {
 		$fields = array(
 			array('name'=>'Name', 	'type'=>'text', 'required'=>true, 'param'=>16, 'extra'=>false, 'visible'=>true),
 			array('name'=>'Description', 	'type'=>'long text', 'required'=>false, 'extra'=>false),
-			array('name'=>'Percentage', 	'type'=>'integer', 'required'=>true, 'extra'=>false, 'visible'=>true)
+			array('name'=>'Percentage', 	'type'=>'float', 'required'=>true, 'extra'=>false, 'visible'=>true)
 		);
 
 		Utils_RecordBrowserCommon::install_new_recordset('data_tax_rates', $fields);
